fix(oci): Better formalizing of (re-)packaging components
This commit fixes an issue where during re-packaging it would
add an additional kernel or an additional initrd if they were
present (i.e. if `ocipack.original` was not `nil`).  To make
this sequence of adding the kernel, the debug kernel and the
initrd a little bit more stable, introduce 3 complementary
methods, respectively, which do just this into the `manifest`
class.  These methods, `SetKernel`, `SetKernelDbg`,
`SetInitrd` iterate through the list of existing layers and
remove the well-known layer with said annotation.

Signed-off-by: Alexander Jung <alex@unikraft.io>